refactor handleVersion, add includeDeaccession param #10240
